Integrated original functionalities into develop
