+++ src/components/review/specimen-view/image-modal.tsx
